Add Table 2.3: Classes of attribute value
This table is referenced in "005-locations-on-stack.md", but does
not exist in our partial "DWARF Specification.txt".

The entry of interest in this table that was referenced by the proposal
is "locdesc". However, it did not exist in DWARFv5 specification. After
digging, I've figured that it only existed in the DWARFv6 DRAFT [1], and
was added by a commit in 2023 [2].

[1] dwarf-spec upstream
https://git.dwarfstd.org/dwarf-spec/

[2] Incorporate approved changes from Oct 16, 2023 meeting.
https://git.dwarfstd.org/dwarf-spec/commit/?id=3663612b002c6b58c486d1693d3b7197ce7a1e1a